refactor(commands): split workspace_cmds.rs (1598 lines) into submodules
U-16 (<=800 lines/file) was violated. Move each cmd_* method, its private
helpers, and the orphan_tests module into dedicated submodules. All
public App::cmd_* signatures unchanged; cmd_workspace_binding_remove
becomes pub(super) so the relocated orphan tests retain their call path.

- workspace_cmds/mod.rs     56 lines
- status.rs                103
- doctor.rs                356  (cmd_doctor + check_projection_drift +
                                  build_doctor_checks +
                                  build_registry_integrity_checks +
                                  doctor_check)
- init.rs                   89
- binding.rs               300
- orphan.rs                695  (cmd_skill_orphan_clean +
                                  cmd_skill_orphan_list + cleanup_*
                                  + is_orphan_projection +
                                  LivePathCleanup + mod orphan_tests)
- remote.rs                 35
- shared.rs                 31
